+++ README.md
@@ -1,0 +1,14 @@
+ABOUT
+
+This project creates config depending upon the environment. 
+It uses convit to add schema for general configuration.
+It also uses dotenv to load environment variables in node.
+
+SETUP
+
+Add .env file at root level with environment variables. 
+Smample: NODE_ENV = development
+
+Change json files inside the server/conf/config as per your need.
+All configurations are loaded in sever.js file
+
