+++ server/conf/config/schema.json
@@ -1,0 +1,28 @@
+{
+    "env": {
+        "doc": "Environment variable",
+        "format": ["development", "production", "test", "staging"],
+        "default": "development",
+        "env": "NODE_ENV"
+    },
+    "port": {
+        "doc": "Port number",
+        "format": "port",
+        "default": 3000,
+        "env": "PORT"
+    },
+    "db": {
+        "name": {
+            "doc": "DB name",
+            "format": "String",
+            "default": "devDB",
+            "env": "DB_NAME"
+        },
+        "host": {
+            "doc": "DB host name",
+            "format": "String",
+            "default": "localhost:27017",
+            "env": "DB_HOST"
+        }
+    }
+}
